refactor: 페이지네이션 컴포넌트 공통화 및 구조 개선
