+++ apps/admin/src/shared/ui/index.ts
@@ -1,1 +1,2 @@
+export { CommonPagination } from './CommonPagination';
 export { default as FormErrorMessage } from './FormErrorMessage';
