+++ apps/admin/src/widgets/students/index.ts
@@ -4,5 +4,4 @@ export * from './model/useUpdateStudent';
 export { default as StudentList } from './ui/StudentList';
 export { default as StudentFilter } from './ui/StudentFilter';
 export { default as StudentFormDialog } from './ui/StudentFormDialog';
-export { default as StudentPagination } from './ui/StudentPagination';
 export { default as StudentExcelActions } from './ui/StudentExcelActions';
